functions: remove office-hours syncOfficeHours + sampleDispatchQueueMetrics; extract github-app-auth (Unit 2 of #2663)
Remove the two office-hours scheduled Cloud Functions. Extract the shared GitHub-App auth helpers (buildAppJwt, mintInstallationToken) into functions/src/github-app-auth.ts so the surviving collectProjectSignals keeps its auth path; repoint project-signals.ts to the new module. Remove the DISPATCH_METRICS_QUEUE_REPO and OFFICE_HOURS_GROUP_REPO references from the deploy workflow. Relocate the KEPT -core test coverage into new *-core.test.ts files and move the auth-helper tests to github-app-auth.test.ts; delete the wrapper test files. feedProxy, collectProjectSignals, all *-core modules, and the shared secrets are kept intact.
