+++ functions/src/index.ts
@@ -1,4 +1,2 @@
 export { feedProxy } from "./feed-proxy.js";
-export { syncOfficeHours } from "./office-hours-sync.js";
-export { sampleDispatchQueueMetrics } from "./dispatch-queue-metrics.js";
 export { collectProjectSignals } from "./project-signals.js";
