+++ functions/test/office-hours-sync-core.test.ts
@@ -1,11 +1,5 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import { createVerify, generateKeyPairSync } from "node:crypto";
 import type { Firestore } from "firebase-admin/firestore";
-
-vi.mock("firebase-admin/app", () => ({
-  getApps: () => [{}],
-  initializeApp: vi.fn(),
-}));
 
 vi.mock("firebase-admin/firestore", () => {
   const TimestampImpl = {
@@ -18,36 +12,12 @@ vi.mock("firebase-admin/firestore", () => {
   };
 });
 
-vi.mock("firebase-functions/v2/scheduler", () => ({
-  onSchedule: vi.fn((_opts, _fn) => ({})),
-}));
-
-vi.mock("firebase-functions/params", () => ({
-  defineSecret: () => ({ value: () => "test-token" }),
-  defineString: (_name: string, opts?: { default?: string }) => ({
-    value: () => opts?.default ?? "test-value",
-  }),
-}));
-
 import {
   syncOfficeHoursCore,
-  fetchOpenJitIssuesLive,
   parseJitDueMarker,
-  buildAppJwt,
-  mintInstallationToken,
   type ReminderItem,
-} from "../src/office-hours-sync";
+} from "../src/office-hours-sync-core";
 import { truncateForLog } from "../src/log-utils";
-
-const testKeyPair = generateKeyPairSync("rsa", {
-  modulusLength: 2048,
-  publicKeyEncoding: { type: "spki", format: "pem" },
-  privateKeyEncoding: { type: "pkcs1", format: "pem" },
-});
-
-function decodeJwtPart(part: string): Record<string, unknown> {
-  return JSON.parse(Buffer.from(part, "base64url").toString("utf8"));
-}
 
 interface InMemoryDocRef {
   path: string;
@@ -430,337 +400,6 @@ describe("parseJitDueMarker", () => {
 
   it("returns null when the marker is malformed", () => {
     expect(parseJitDueMarker("<!-- jit-due: not-a-date -->")).toBeNull();
-  });
-});
-
-describe("fetchOpenJitIssuesLive", () => {
-  beforeEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it("posts a GraphQL query and returns jit-labelled issues with parsed due times", async () => {
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: true,
-      status: 200,
-      text: () => Promise.resolve(""),
-      json: () =>
-        Promise.resolve({
-          data: {
-            repository: {
-              issues: {
-                pageInfo: { endCursor: null, hasNextPage: false },
-                nodes: [
-                  {
-                    number: 42,
-                    title: "Daily chore",
-                    body:
-                      "Body text\n\n<!-- jit-due: 2026-01-15T12:00:00Z -->",
-                    labels: {
-                      nodes: [{ name: "jit:daily-chore" }, { name: "help wanted" }],
-                    },
-                  },
-                  {
-                    number: 50,
-                    title: "Old jit issue without a marker",
-                    body: "Just the body, no marker yet.",
-                    labels: { nodes: [{ name: "jit:legacy" }] },
-                  },
-                  {
-                    number: 43,
-                    title: "Non-jit issue",
-                    body: "Body",
-                    labels: { nodes: [{ name: "bug" }] },
-                  },
-                ],
-              },
-            },
-          },
-        }),
-    });
-    vi.stubGlobal("fetch", fetchMock);
-
-    const issues = await fetchOpenJitIssuesLive("natb1/office-hours-nate", "test-token");
-
-    expect(fetchMock).toHaveBeenCalledTimes(1);
-    const [url, init] = fetchMock.mock.calls[0] as [string, RequestInit];
-    expect(url).toBe("https://api.github.com/graphql");
-    expect(init.method).toBe("POST");
-    const headers = init.headers as Record<string, string>;
-    expect(headers.Authorization).toBe("bearer test-token");
-    expect(headers["User-Agent"]).toBe("office-hours-sync/1.0");
-    expect(headers["Content-Type"]).toBe("application/json");
-    const body = JSON.parse(init.body as string) as {
-      query: string;
-      variables: { owner: string; name: string; cursor: string | null };
-    };
-    expect(body.query).toContain("repository(owner: $owner, name: $name)");
-    expect(body.variables).toEqual({
-      owner: "natb1",
-      name: "office-hours-nate",
-      cursor: null,
-    });
-
-    expect(issues).toHaveLength(2);
-
-    const reminder0 = issues[0] as ReminderItem;
-    expect(reminder0).toMatchObject({
-      kind: "reminder",
-      number: 42,
-      title: "Daily chore",
-      jitKey: "daily-chore",
-      repo: "natb1/office-hours-nate",
-    });
-    expect(reminder0.dueAt?.toISOString()).toBe("2026-01-15T12:00:00.000Z");
-
-    const reminder1 = issues[1] as ReminderItem;
-    expect(reminder1).toMatchObject({
-      kind: "reminder",
-      number: 50,
-      jitKey: "legacy",
-      dueAt: null,
-    });
-  });
-
-  it("throws when the fetch returns non-OK", async () => {
-    vi.stubGlobal(
-      "fetch",
-      vi.fn().mockResolvedValue({
-        ok: false,
-        status: 500,
-        text: () => Promise.resolve("server error"),
-        json: () => Promise.resolve({}),
-      }),
-    );
-
-    await expect(
-      fetchOpenJitIssuesLive("natb1/office-hours-nate", "test-token"),
-    ).rejects.toThrow(/500/);
-  });
-
-  it("throws when the GraphQL response carries errors", async () => {
-    vi.stubGlobal(
-      "fetch",
-      vi.fn().mockResolvedValue({
-        ok: true,
-        status: 200,
-        text: () => Promise.resolve(""),
-        json: () =>
-          Promise.resolve({
-            errors: [{ message: "Bad credentials" }],
-          }),
-      }),
-    );
-
-    await expect(
-      fetchOpenJitIssuesLive("natb1/office-hours-nate", "test-token"),
-    ).rejects.toThrow(/Bad credentials/);
-  });
-
-  it("follows pagination and collects issues from multiple pages", async () => {
-    const page1 = {
-      data: {
-        repository: {
-          issues: {
-            pageInfo: { endCursor: "cursor-abc", hasNextPage: true },
-            nodes: [
-              {
-                number: 1,
-                title: "Page 1 issue",
-                body: "<!-- jit-due: 2026-01-15T12:00:00Z -->",
-                labels: { nodes: [{ name: "jit:page1-key" }] },
-              },
-            ],
-          },
-        },
-      },
-    };
-    const page2 = {
-      data: {
-        repository: {
-          issues: {
-            pageInfo: { endCursor: null, hasNextPage: false },
-            nodes: [
-              {
-                number: 2,
-                title: "Page 2 issue",
-                body: "<!-- jit-due: 2026-02-01T00:00:00Z -->",
-                labels: { nodes: [{ name: "jit:page2-key" }] },
-              },
-            ],
-          },
-        },
-      },
-    };
-
-    const fetchMock = vi
-      .fn()
-      .mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        text: () => Promise.resolve(""),
-        json: () => Promise.resolve(page1),
-      })
-      .mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        text: () => Promise.resolve(""),
-        json: () => Promise.resolve(page2),
-      });
-    vi.stubGlobal("fetch", fetchMock);
-
-    const issues = await fetchOpenJitIssuesLive("natb1/office-hours-nate", "test-token");
-
-    expect(fetchMock).toHaveBeenCalledTimes(2);
-    // Second call must forward the cursor from the first page.
-    const secondCall = fetchMock.mock.calls[1] as [string, RequestInit];
-    const secondBody = JSON.parse(secondCall[1].body as string) as {
-      variables: { cursor: string | null };
-    };
-    expect(secondBody.variables.cursor).toBe("cursor-abc");
-
-    expect(issues).toHaveLength(2);
-    expect((issues[0] as ReminderItem).jitKey).toBe("page1-key");
-    expect((issues[1] as ReminderItem).jitKey).toBe("page2-key");
-  });
-
-  it("stops pagination when hasNextPage is true but endCursor is null", async () => {
-    // A malformed GitHub response with hasNextPage=true but null endCursor should
-    // not loop infinitely — the guard should stop after the first page.
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: true,
-      status: 200,
-      text: () => Promise.resolve(""),
-      json: () =>
-        Promise.resolve({
-          data: {
-            repository: {
-              issues: {
-                pageInfo: { endCursor: null, hasNextPage: true },
-                nodes: [
-                  {
-                    number: 10,
-                    title: "Issue on only page",
-                    body: "<!-- jit-due: 2026-03-01T00:00:00Z -->",
-                    labels: { nodes: [{ name: "jit:solo-key" }] },
-                  },
-                ],
-              },
-            },
-          },
-        }),
-    });
-    vi.stubGlobal("fetch", fetchMock);
-
-    const issues = await fetchOpenJitIssuesLive("natb1/office-hours-nate", "test-token");
-
-    // Must not loop — one fetch only.
-    expect(fetchMock).toHaveBeenCalledTimes(1);
-    expect(issues).toHaveLength(1);
-    expect((issues[0] as ReminderItem).jitKey).toBe("solo-key");
-  });
-});
-
-describe("buildAppJwt", () => {
-  it("produces an RS256 JWT signed by the App private key with the expected claims", () => {
-    const nowMs = 1_700_000_000_000;
-    const jwt = buildAppJwt("123456", testKeyPair.privateKey, nowMs);
-
-    const [header, payload, signature] = jwt.split(".");
-    expect(header).toBeTruthy();
-    expect(payload).toBeTruthy();
-    expect(signature).toBeTruthy();
-
-    expect(decodeJwtPart(header)).toEqual({ alg: "RS256", typ: "JWT" });
-
-    const claims = decodeJwtPart(payload) as {
-      iat: number;
-      exp: number;
-      iss: string;
-    };
-    const expectedIat = Math.floor(nowMs / 1000) - 60;
-    expect(claims.iss).toBe("123456");
-    expect(claims.iat).toBe(expectedIat);
-    expect(claims.exp).toBe(expectedIat + 9 * 60);
-
-    const verified = createVerify("RSA-SHA256")
-      .update(`${header}.${payload}`)
-      .verify(testKeyPair.publicKey, Buffer.from(signature, "base64url"));
-    expect(verified).toBe(true);
-  });
-});
-
-describe("mintInstallationToken", () => {
-  beforeEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it("exchanges an App JWT for an installation token", async () => {
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: true,
-      status: 201,
-      text: () => Promise.resolve(""),
-      json: () => Promise.resolve({ token: "ghs_installationtoken" }),
-    });
-    vi.stubGlobal("fetch", fetchMock);
-
-    const token = await mintInstallationToken({
-      appId: "123456",
-      installationId: "87654321",
-      privateKey: testKeyPair.privateKey,
-      nowMs: 1_700_000_000_000,
-    });
-
-    expect(token).toBe("ghs_installationtoken");
-    expect(fetchMock).toHaveBeenCalledTimes(1);
-    const [url, init] = fetchMock.mock.calls[0] as [string, RequestInit];
-    expect(url).toBe(
-      "https://api.github.com/app/installations/87654321/access_tokens",
-    );
-    expect(init.method).toBe("POST");
-    const headers = init.headers as Record<string, string>;
-    expect(headers.Authorization).toMatch(/^Bearer .+\..+\..+$/);
-    expect(headers.Accept).toBe("application/vnd.github+json");
-    expect(headers["User-Agent"]).toBe("office-hours-sync/1.0");
-  });
-
-  it("throws when the token exchange returns non-OK", async () => {
-    vi.stubGlobal(
-      "fetch",
-      vi.fn().mockResolvedValue({
-        ok: false,
-        status: 401,
-        text: () => Promise.resolve("Bad credentials"),
-        json: () => Promise.resolve({}),
-      }),
-    );
-
-    await expect(
-      mintInstallationToken({
-        appId: "123456",
-        installationId: "87654321",
-        privateKey: testKeyPair.privateKey,
-      }),
-    ).rejects.toThrow(/401/);
-  });
-
-  it("throws when the token exchange response omits the token", async () => {
-    vi.stubGlobal(
-      "fetch",
-      vi.fn().mockResolvedValue({
-        ok: true,
-        status: 201,
-        text: () => Promise.resolve(""),
-        json: () => Promise.resolve({}),
-      }),
-    );
-
-    await expect(
-      mintInstallationToken({
-        appId: "123456",
-        installationId: "87654321",
-        privateKey: testKeyPair.privateKey,
-      }),
-    ).rejects.toThrow(/missing token/);
   });
 });
